requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
lxml==5.3.0
MarkupSafe==2.1.5
openai==1.42.0
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pytesseract==0.3.13
python-dotenv==1.0.1
python-pptx==1.0.2
requests==2.32.3
sniffio==1.3.1
tesseract==0.1.3
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
XlsxWriter==3.2.0
gunicorn==20.1.0
google-cloud-storage==2.18.2